requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.18.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
pefile==2019.4.18
PyInstaller==3.6
PySimpleGUI==4.20.0
python-firebase==1.2
pywin32-ctypes==0.2.0
requests==2.24.0
urllib3==1.25.9
Werkzeug==1.0.1
